[#13] Add mockup of project upload interface
A very crude mockup. Missing is image upload widgets. In their simplest
form they would be similar to the XML files upload, and would upload a
zip file with all images. This might do to start with, but it means
unpacking the zip file on the recieveing end and display of the result
(if any) is an unsolved part.

Also included is the XML export from Balsamiq for the mockup.